@@ -3,8 +3,8 @@ using CodeBreaker.Blazor.Resources;
 using CodeBreaker.Blazor.ViewModels;
 using CodeBreaker.Services;
 using CodeBreaker.Shared.Models.Api;
-using CodeBreaker.UI.Shared.Models.DataGrid;
-using CodeBreaker.UI.Shared.Services.Dialog;
+using CodeBreaker.UI.Models.DataGrid;
+using CodeBreaker.UI.Services.Dialog;
 using Microsoft.AspNetCore.Components;
 using Microsoft.Extensions.Localization;
 
@@ -13,45 +13,44 @@ namespace CodeBreaker.Blazor.Pages;
 public partial class ReportsPage
 {
     [Inject]
-    private ICodeBreakerDialogService _codeBreakerDialogService { get; set; } = default!;
+    private IDialogService DialogService { get; set; } = default!;
 
     [Inject]
-    private IGameReportClient _gameClient { get; set; } = default!;
+    private IGameReportClient GameClient { get; set; } = default!;
 
     [Inject]
-    private ILogger<ReportsPage> _logger { get; set; } = default!;
+    private ILogger<ReportsPage> Logger { get; set; } = default!;
     [Inject]
     private IStringLocalizer<Resource> Loc { get; set; } = default!;
 
 
-    private List<GameDto> _games = new();
+    private List<GameDto> _games = [];
     private bool _isLoadingGames = false;
-    private ReportFilterContext _filter = new();
+    private readonly ReportFilterContext _filter = new();
     
-    private List<string> _headers => Loc.GetString("Reports_Table_Headers").Value.Split(",").ToList();
+    private List<string> Headers => [.. Loc.GetString("Reports_Table_Headers").Value.Split(",")];
 
-    private List<CodeBreakerColumnDefinition<GameDto>> _columns = new()
-    {
+    private readonly List<CodeBreakerColumnDefinition<GameDto>> _columns = [
         new CodeBreakerColumnDefinition<GameDto>("Gamername", game => game.Username, true),
         new CodeBreakerColumnDefinition<GameDto>("Start", game => game.Start, false),
         new CodeBreakerColumnDefinition<GameDto>("End", game => game.End.HasValue ? game.End.Value : "----", false),
         new CodeBreakerColumnDefinition<GameDto>("Number of Moves", game => game.Moves.Count(), true)
-    };
+    ];
 
     public async Task GetGames()
     {
-        _logger?.LogInformation("Calling GetReport for {date}", _filter.Date);
-        _games = new List<GameDto>();
+        Logger?.LogInformation("Calling GetReport for {date}", _filter.Date);
+        _games = [];
         _isLoadingGames = true;
         try
         {
-            GetGamesResponse? response = await _gameClient.GetGamesAsync(_filter.Date);
-            _logger?.LogDebug("Got response", response);
-            _games = response?.Games?.ToList() ?? new List<GameDto>();
+            GetGamesResponse? response = await GameClient.GetGamesAsync(_filter.Date);
+            Logger?.LogDebug("Got response", response);
+            _games = [..response?.Games ?? []];
         }
         catch (Exception ex)
         {
-            _logger?.LogError(ex, "Error calling GetGames");
+            Logger?.LogError(ex, "Error calling GetGames");
             //TODO: handle Exception;
         }
         finally
@@ -80,10 +79,10 @@ public partial class ReportsPage
             }
         }
 
-        _codeBreakerDialogService.ShowDialog(new CodeBreakerDialogContext(typeof(Playground), new Dictionary<string, object>
+        DialogService.ShowDialog(new DialogContext(typeof(Playground), new Dictionary<string, object>
             {
                 { nameof(Playground.Game), game },
                 { nameof(Playground.GameFinished), true },
-            }, title, new List<CodeBreakerDialogActionContext>()));
+            }, title, []));
     }
 }
