@@ -2,23 +2,22 @@ using Microsoft.AspNetCore.Components;
 using Microsoft.AspNetCore.Components.WebAssembly.Authentication;
 using Microsoft.Extensions.Options;
 
-namespace CodeBreaker.Blazor.Shared
+namespace CodeBreaker.Blazor.Shared;
+
+public partial class LoginDisplay
 {
-    public partial class LoginDisplay
+    [Inject] private NavigationManager NavigationManager { get; set; } = default!;
+    [Inject] private IOptionsSnapshot<RemoteAuthenticationOptions<ApiAuthorizationProviderOptions>> Options { get; set; } = default!;
+
+    private void BeginLogOut()
     {
-        [Inject] private NavigationManager _navigationManager { get; set; } = default!; 
-        [Inject] private IOptionsSnapshot<RemoteAuthenticationOptions<ApiAuthorizationProviderOptions>> _options { get; set; } = default!;
+        NavigationManager.NavigateToLogout(Options.Get(Microsoft.Extensions.Options.Options.DefaultName)
+            .AuthenticationPaths.LogOutPath);
+    }
 
-        private void BeginLogOut()
-        {
-            _navigationManager.NavigateToLogout(_options.Get(Options.DefaultName)
-                .AuthenticationPaths.LogOutPath);
-        }
-
-        private void BeginLogIn()
-        {
-            _navigationManager.NavigateToLogin(_options.Get(Options.DefaultName)
-                .AuthenticationPaths.LogInPath);
-        }
+    private void BeginLogIn()
+    {
+        NavigationManager.NavigateToLogin(Options.Get(Microsoft.Extensions.Options.Options.DefaultName)
+            .AuthenticationPaths.LogInPath);
     }
 }
