@@ -1,15 +1,30 @@
-using Microsoft.AspNetCore.ResponseCompression;
+using CodeBreaker.Blazor.Host.Components;
+using CodeBreaker.Blazor.Pages;
+using CodeBreaker.Services.Authentication;
+using CodeBreaker.Services;
+using CodeBreaker.UI;
+using CodeBreaker.UI.Services.Dialog;
+using CodeBreaker.Blazor.Host.Services;
 
 var builder = WebApplication.CreateBuilder(args);
 
-// Add services to the container.
+builder.Services.AddRazorComponents()
+    .AddInteractiveServerComponents()
+    .AddInteractiveWebAssemblyComponents();
 
-builder.Services.AddControllersWithViews();
-builder.Services.AddRazorPages();
+builder.Services.AddLocalization();
+builder.Services.AddCodeBreakerUI();
+
+builder.Services.AddHttpClient("GameApi", configure =>
+    configure.BaseAddress = new Uri(builder.Configuration.GetRequired("ApiBase")));
+
+builder.Services.AddScoped<IAuthService, DummyAuthService>();
+builder.Services.AddHttpClient<IGameClient, GameClient>("GameApi");
+builder.Services.AddHttpClient<IGameReportClient, GameClient>("GameApi");
+builder.Services.AddScoped<IDialogService, DialogService>();
 
 var app = builder.Build();
 
-// Configure the HTTP request pipeline.
 if (app.Environment.IsDevelopment())
 {
     app.UseWebAssemblyDebugging();
@@ -22,15 +37,15 @@ else
 }
 
 app.UseHttpsRedirection();
-
-app.UseBlazorFrameworkFiles();
 app.UseStaticFiles();
+app.UseAntiforgery();
+app.UseRequestLocalization(new RequestLocalizationOptions()
+    .AddSupportedCultures(["de", "en"])
+    .AddSupportedUICultures(["de", "en"]));
 
-app.UseRouting();
-
-
-app.MapRazorPages();
-app.MapControllers();
-app.MapFallbackToFile("index.html");
+app.MapRazorComponents<App>()
+    .AddInteractiveServerRenderMode()
+    .AddInteractiveWebAssemblyRenderMode()
+    .AddAdditionalAssemblies(typeof(GamePage).Assembly);
 
 app.Run();
