@@ -1,6 +1,6 @@
 ﻿@using System.Globalization
 
-@if (_items.Count > 0)
+@if (_items.Count != 0)
 {
-    <CodeBreakerSelect Items="_items" @bind-Value="Culture"></CodeBreakerSelect>
+    <CodeBreakerSelect Items="_items" @bind-Value="Culture" />
 }
