@@ -10,11 +10,11 @@
                 <p class="content">@version</p>
                 <p class=title>@Loc["About_Information_Theme"]</p>
                 <div class="content">
-                    <CodeBreakerThemeSwitch></CodeBreakerThemeSwitch>
+                    <CodeBreakerThemeSwitch />
                 </div>
                 <p class="title">@Loc["Language_Select_Label"]</p>
                 <div class="content">
-                    <LanguageSelector></LanguageSelector>
+                    <LanguageSelector />
                 </div>
             </div>
         </div>
