@@ -1,16 +1,15 @@
 ﻿@page "/authentication/{action}"
 @using Microsoft.AspNetCore.Components.WebAssembly.Authentication
 
+@*
 <RemoteAuthenticatorView Action="@Action" OnLogOutSucceeded="OnLogout">
-    <LoggingIn></LoggingIn>
-    <CompletingLoggingIn></CompletingLoggingIn>
-    <CompletingLogOut></CompletingLogOut>
-    <LoggingIn></LoggingIn>
-    <LogInFailed></LogInFailed>
-    <LogOut></LogOut>
-    <LogOutFailed></LogOutFailed>
-    <LogOutSucceeded>
-    </LogOutSucceeded>
+    <LoggingIn />
+    <CompletingLoggingIn />
+    <CompletingLogOut />
+    <LoggingIn />
+    <LogInFailed />
+    <LogOut />
+    <LogOutFailed />
+    <LogOutSucceeded />
 </RemoteAuthenticatorView>
-
-
+ *@
