@@ -1,7 +1,7 @@
 ﻿@page "/"
 @inject IStringLocalizer<Resource> Loc
 
-<PageTitle>Index</PageTitle>
+<PageTitle>@Loc["AppDisplayName"]</PageTitle>
 
 <h1 class="page-title">@Loc["Index_Welcome"] <b>@Loc["AppDisplayName"]</b></h1>
 
