@@ -1,15 +1,16 @@
-﻿@using UI.Shared.Models.Icon
+﻿@using UI.Models.Icon
 
-<AuthorizeView>
+@* <AuthorizeView>
     <Authorized>
         <div class="user">
-            @if (!String.IsNullOrWhiteSpace(context.User.Identity?.Name)) {
+            @if (!string.IsNullOrWhiteSpace(context.User.Identity?.Name)) {
                 <a href="/user-information">@context.User.Identity.Name</a>
             }
-            <CodeBreakerIconButton OnClick="BeginLogOut" Icon="CodeBreakerIcon.Logout"></CodeBreakerIconButton>
+            <CodeBreakerIconButton OnClick="BeginLogOut" Icon="CodeBreakerIcon.Logout" />
         </div>
     </Authorized>
     <NotAuthorized>
-        <CodeBreakerIconButton OnClick="BeginLogIn" Icon="CodeBreakerIcon.Login"></CodeBreakerIconButton>
+        <CodeBreakerIconButton OnClick="BeginLogIn" Icon="CodeBreakerIcon.Login" />
     </NotAuthorized>
 </AuthorizeView>
+ *@
