@@ -1,16 +1,13 @@
 ﻿@inherits LayoutComponentBase
-@inject IStringLocalizer<Resource> Loc
 
 <CodeBreakerLayout Label="@($"{Loc["Index_Welcome"]} {Loc["AppDisplayName"]}")">
     <ToolbarContent>
-        <div class="flex toolbar-content">
-            <LoginDisplay></LoginDisplay>
-        </div>
+        <div class="flex toolbar-content"></div>
     </ToolbarContent>
     <ChildContent>
         <div class="content-container">@Body</div>
     </ChildContent>
     <DrawerContent>
-        <CodeBreakerNavMenu MenuItems="_menuItems"></CodeBreakerNavMenu>
+        <CodeBreakerNavMenu MenuItems="_menuItems" />
     </DrawerContent>
 </CodeBreakerLayout>
