@@ -1,6 +1,6 @@
 ﻿@using System.Net.Http
 @using System.Net.Http.Json
-@using Microsoft.AspNetCore.Components.Authorization
+@* @using Microsoft.AspNetCore.Components.Authorization *@
 @using Microsoft.AspNetCore.Components.Forms
 @using Microsoft.AspNetCore.Components.Routing
 @using Microsoft.AspNetCore.Components.Web
@@ -14,3 +14,5 @@
 @using CodeBreaker.Blazor.Resources
 @using CodeBreaker.UI
 @using BlazorApplicationInsights
+@using static Microsoft.AspNetCore.Components.Web.RenderMode
+
